# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateExpenseRequest:
    properties:
      amount:
        description: CategoryID  uint      `json:"category_id" validate:"required"`
        example: 25.5
        type: number
      date:
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        maxLength: 500
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - amount
    - date
    type: object
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        example: Validation failed
        type: string
      message:
        example: Email is required
        type: string
    type: object
  dto.ExpenseAnalytics:
    properties:
      average_expense_amount:
        type: number
      average_per_day:
        type: number
      expenses_count:
        type: integer
      largest_expense:
        $ref: '#/definitions/dto.ExpenseResponse'
      period:
        type: string
      smallest_expense:
        $ref: '#/definitions/dto.ExpenseResponse'
      total_amount:
        type: number
    type: object
  dto.ExpensePeriod:
    properties:
      period:
        type: string
    type: object
  dto.ExpenseResponse:
    properties:
      amount:
        description: Category     CategoryResponse `json:"category,omitempty"`
        type: number
      category_id:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
  dto.ExpensesListResponse:
    properties:
      expenses:
        items:
          $ref: '#/definitions/dto.ExpenseResponse'
        type: array
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Система управления личными финансами
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Finance API
  version: "1.0"
paths:
...
  /categories/{category_id}/expenses:
    get:
      consumes:
      - application/json
      description: Получение всех расходов пользователя в указанной категории
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список расходов
          schema:
            $ref: '#/definitions/dto.ExpensesListResponse'
        "400":
          description: Неверный ID категории
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка расходов
      tags:
      - Expenses
    post:
      consumes:
      - application/json
      description: Создание нового расхода в указанной категории
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      - description: Данные для создания расхода
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Расход успешно создан
          schema:
            $ref: '#/definitions/dto.ExpenseResponse'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание нового расхода
      tags:
      - Expenses
  /categories/{category_id}/expenses/{expense_id}:
    delete:
      consumes:
      - application/json
      description: Удаление конкретного расхода пользователя
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      - description: ID расхода
        in: path
        name: expense_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Расход успешно удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID категории или расхода
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление расхода
      tags:
      - Expenses
    get:
      consumes:
      - application/json
      description: Получение информации о конкретном расходе
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      - description: ID расхода
        in: path
        name: expense_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о расходе
          schema:
            $ref: '#/definitions/dto.ExpenseResponse'
        "400":
          description: Неверный ID категории или расхода
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение расхода по ID
      tags:
      - Expenses
  /categories/{category_id}/expenses/analytics:
    post:
      consumes:
      - application/json
      description: Получение аналитики расходов по категории за указанный период
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      - description: Период для анализа
        in: body
        name: period
        required: true
        schema:
          $ref: '#/definitions/dto.ExpensePeriod'
      produces:
      - application/json
      responses:
        "200":
          description: Аналитика расходов
          schema:
            $ref: '#/definitions/dto.ExpenseAnalytics'
        "400":
          description: Неверный ID категории или период
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение аналитики расходов
      tags:
      - Expenses
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"